Assignments/02._Data/data_parsing_server/pyproject.toml
[project]
name = "data-parsing-server"
version = "0.1.0"
description = ""
authors = [
    {name = "VictorHanert",email = "vict660m@stud.kea.dk"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "xmltodict (>=0.14.2,<0.15.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
